+++ utils.go
@@ -1,9 +1,9 @@
 package bitvec
 
-func word(i int) int {
+func word(i uint64) uint64 {
 	return i / wordSize
 }
 
-func offset(i int) uint {
-	return uint(i % wordSize)
+func offset(i uint64) uint64 {
+	return uint64(i % wordSize)
 }
